+++ src/RssItem.ts
@@ -1,0 +1,7 @@
+export interface RssItem {
+    title?: string;
+    pubDate?: string;
+    image?: string;
+    description?: string;
+    link?: string;
+}
